feat: don't overwrite existing service

+++ generator_scripts/generate_python_service.sh
@@ -1,4 +1,5 @@
 service=$1_py
+if [ -d services/$service ]; then echo "error: service already exists" >>/dev/stderr; exit 1; fi;
 mkdir services/$service
 echo "ping.proto" > services/$service/schemas.txt
 echo "protobuf==3.17.3" > services/$service/requirements.txt
